hot fix at Allure.step at senKeysInput
